feat: add a flag to the init command to allow it to run without user input
