+++ scripts/generate_docs.py
@@ -6,7 +6,7 @@ import sys
 
 from fides.api.main import app
 from fides.core.config import CONFIG
-from fides.core.config.docs import generate_config_docs
+from fides.core.config.create import generate_config_docs
 
 
 def generate_openapi(outfile_dir: str) -> str:
